3.1.1 przeniesienie implementacji stosu do nowej klasy IntArrayStack
